move testing functionality to own header
