+++ src/lib/tests/benchmark.cpp
@@ -1,33 +1,13 @@
-#include "distancekernels.h"
-#include "distopia.h"
-#include "vanilla.h"
 #include <benchmark/benchmark.h>
 #include <iostream>
 #include <random>
 
+#include "distancekernels.h"
+#include "distopia.h"
+#include "vanilla.h"
+#include "test_utils.h"
+
 #define BOXSIZE 30
-
-// creates nrandom floating points between 0 and limit
-template <typename T>
-void RandomFloatingPoint(T *target, const int nrandom, const int neglimit,
-                         const int poslimit) {
-  std::random_device rd;
-  std::mt19937 gen(rd()); // Standard mersenne_twister_engine
-  std::uniform_real_distribution<T> distribution(neglimit, poslimit);
-  for (size_t i = 0; i < nrandom; i++) {
-    target[i] = distribution(gen);
-  }
-}
-
-void RandomInt(std::size_t *target, const int nrandom, const int neglimit,
-               const int poslimit) {
-  std::random_device rd;
-  std::mt19937 gen(rd()); // Standard mersenne_twister_engine
-  std::uniform_int_distribution<std::size_t> distribution(neglimit, poslimit);
-  for (size_t i = 0; i < nrandom; i++) {
-    target[i] = distribution(gen);
-  }
-}
 
 template <typename T> class CoordinatesDynamicMem : public benchmark::Fixture {
 public:
